# lib/api-zod/src/generated/types/highlight.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kinectem API - social platform for youth athletes
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface Highlight {
  id: string;
  title: string;
  teamId: string;
  teamName?: string;
  articleId?: string;
  articleTitle?: string;
  thumbnailUrl?: string;
  videoUrl?: string;
  durationSeconds?: number;
  createdAt?: Date;
  taggedUsers?: User[];
}