# lib/api-zod/src/generated/types/newArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kinectem API - social platform for youth athletes
 * OpenAPI spec version: 0.1.0
 */

export interface NewArticle {
  title: string;
  teamId: string;
  opponentName?: string;
  gameDate?: Date;
  gameScore?: string;
  snippet?: string;
  body: string;
  coverImageUrl?: string;
  taggedUserIds?: string[];
  highlightIds?: string[];
}